% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{clean function}
\usage{
clean(
  df,
  column,
  replacement = "",
  selected = NULL,
  prefixes = FALSE,
  suffixes = FALSE,
  switch_order = FALSE,
  ignore.case = FALSE
)
}
\arguments{
\item{df}{A table the column is in}

\item{column}{The column to be cleaned}

\item{replacement}{What the cleaned portion of elements will be changed to, Default: ''}

\item{selected}{Characters to removed, Default: NULL}

\item{prefixes}{If true, prefixes will be removed, Default: FALSE}

\item{suffixes}{If true, suffixes will be removed, Default: FALSE}

\item{switch_order}{Should be set to True if one of the columns is
Last Name First and the other is not. Will change the order of names.
Ignore if you not joining by names, Default: FALSE}

\item{ignore.case}{If true, will ignore capitalization, Default: FALSE}
}
\value{
Will return a table which has specified characters removed from the column
}
\description{
Function to clean the column to be used in joining.
Will remove certain parts of elements in a column
}
\details{
This should be run prior to fuzzy_match as it will ensure a higher degree of accuracy.
}
\examples{
\dontrun{
congress <- clean(congress, name, selected = ",", prefixes = T, suffixes = T)
politwoops <- clean(politwoops, full_name, selected = ",", prefixes = T, suffixes = T)
}
}
\seealso{
\code{\link[rapportools]{is.boolean}}
 \code{\link[varhandle]{unfactor}}
}
