% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/committees.R
\docType{data}
\name{committees}
\alias{committees}
\title{Committees Dataframe}
\format{A data frame with 5748 rows and 9 variables:
\describe{
  \item{\code{name}}{character Committee Member's Name}
  \item{\code{party}}{character Committee Member's Pary}
  \item{\code{rank}}{integer Member's rank in the committee}
  \item{\code{title}}{character Leadership position, if any}
  \item{\code{house_committee_id}}{character Committee ID}
  \item{\code{jurisdiction}}{character Jursidiction}
  \item{\code{committee_name}}{character Committee Name}
  \item{\code{sub_name}}{character Sub-Committee Name}
  \item{\code{type}}{integer type of committee: house, senate or joint}
}}
\source{
\href{https://data.world/govtrack/congressional-committees}{https://data.world/govtrack/congressional-committees}
}
\usage{
committees
}
\description{
Data on the Committees of the US Congress
}
\details{
Data on the Committees of the US Congress
}
\keyword{datasets}
