% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuz.merge-package.R
\docType{package}
\name{fuz.merge-package}
\alias{fuz.merge}
\alias{fuz.merge-package}
\title{fuz.merge: This package provides functions to implement and aid in Fuzzy Matching}
\description{
This package lets you implement fuzzy merging so you can merge datasets
    when the values to be matched aren't exactly the same. It also includes a clean
    function to remove prefixes, sufixes, and other sequences of characters which
    the user wants to remove. It also includes a function, scramble(), which will change
    some of the characters of the selected column to allow for testing of the package.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/kstevenson97/fuz.merge/issues}
}

}
\author{
\strong{Maintainer}: Kenneth Stevenson \email{kstevenson@wesleyan.edu}

Authors:
\itemize{
  \item Harrison Karp \email{hkarp@wesleyan.edu}
}

}
\keyword{internal}
