% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scramble.R
\name{scramble}
\alias{scramble}
\title{Scramble Function}
\usage{
scramble(col, chars)
}
\arguments{
\item{col}{Data frame column or list to change}

\item{chars}{Number of characters to change}
}
\value{
New list with some of the characters of random elements changed
}
\description{
Function to change a selected number of characters for elements in a list
}
\details{
Function will change the inputed number of characters from a list or column
}
\examples{
\dontrun{
congress$name <- scramble(congress$name, 2)
 }
}
